Update and rename match_ticker.lua to match_ticker_display.lua
